Split up digits and divmod package
