+++ digits/divmod_test.go
@@ -1,0 +1,12 @@
+package digits_test
+
+import (
+	. "github.com/onsi/ginkgo/v2"
+	. "github.com/onsi/gomega"
+	"testing"
+)
+
+func TestSuite(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "digits package Test Suite")
+}
