+++ divmod/add.go
@@ -1,7 +1,9 @@
 package divmod
 
+import "github.com/borisskert/go-biginteger/digits"
+
 func Add(a, b []uint64) []uint64 {
-	result := Wrap(a).Add(Wrap(b))
+	result := digits.Wrap(a).Add(digits.Wrap(b))
 	result.NormalizeInPlace()
 
 	return result.Trim().Array()
