+++ divmod/knuth.go
@@ -1,26 +1,28 @@
 package divmod
 
-func divModOptimized(numerator Digits, denominator Digits) (Digits, Digits) {
-	if denominator.Compare(ZeroAsDigits()) == 0 {
+import "github.com/borisskert/go-biginteger/digits"
+
+func divModOptimized(numerator digits.Digits, denominator digits.Digits) (digits.Digits, digits.Digits) {
+	if denominator.Compare(digits.ZeroAsDigits()) == 0 {
 		panic("Division by zero")
 	}
 
-	if numerator.Compare(ZeroAsDigits()) == 0 {
-		return ZeroAsDigits(), ZeroAsDigits()
+	if numerator.Compare(digits.ZeroAsDigits()) == 0 {
+		return digits.ZeroAsDigits(), digits.ZeroAsDigits()
 	}
 
-	if denominator.Compare(One()) == 0 {
-		return numerator, ZeroAsDigits()
+	if denominator.Compare(digits.One()) == 0 {
+		return numerator, digits.ZeroAsDigits()
 	}
 
 	cmp := numerator.Compare(denominator)
 
 	if cmp < 0 {
-		return ZeroAsDigits(), numerator
+		return digits.ZeroAsDigits(), numerator
 	}
 
 	if cmp == 0 {
-		return One(), ZeroAsDigits()
+		return digits.One(), digits.ZeroAsDigits()
 	}
 
 	length := numerator.Length()
@@ -42,20 +44,20 @@ func divModOptimized(numerator Digits, denominator Digits) (Digits, Digits) {
 	return divModByDonaldKnuthsTAOCPv2(numerator.Trim(), denominator.Trim())
 }
 
-func d1Normalize(n uint64, b DoubleDigit, u Digits, v Digits) (Digits, Digits, DoubleDigit) {
-	var d DoubleDigit
+func d1Normalize(n uint64, b digits.DoubleDigit, u digits.Digits, v digits.Digits) (digits.Digits, digits.Digits, digits.DoubleDigit) {
+	var d digits.DoubleDigit
 
-	bMaxValue, _ := b.SubtractDigit(One().AsDigit()) // b-1
-	vLast := v.DigitAt(uint(n - 1))                  // Most significant digit of V
+	bMaxValue, _ := b.SubtractDigit(digits.One().AsDigit()) // b-1
+	vLast := v.DigitAt(uint(n - 1))                         // Most significant digit of V
 
 	// Let d = ⌊b / (v_{n-1} + 1)⌋
 
 	if vLast.IsEqualDoubleDigit(bMaxValue) {
 		// If the most significant digit is b-1, set d to 1
-		d = One().AsDoubleDigit()
+		d = digits.One().AsDoubleDigit()
 	} else {
 		// Otherwise, calculate d
-		vLastPlus1, _ := vLast.AsDoubleDigit().AddDigit(OneAsDigit())
+		vLastPlus1, _ := vLast.AsDoubleDigit().AddDigit(digits.OneAsDigit())
 		d, _ = b.Divide128(vLastPlus1) // Divide b by v_{n-1}+1 to get the scaling factor
 	}
 
@@ -66,7 +68,7 @@ func d1Normalize(n uint64, b DoubleDigit, u Digits, v Digits) (Digits, Digits, D
 	return r, v, d
 }
 
-func d3ACalculateQHat(u Digits, v Digits, j int64, n int64) (DoubleDigit, Digit) {
+func d3ACalculateQHat(u digits.Digits, v digits.Digits, j int64, n int64) (digits.DoubleDigit, digits.Digit) {
 	ujPlusN := u.DigitAt(uint(j + n))
 	uJPlusNMinus1 := u.DigitAt(uint(j + n - 1))
 
@@ -82,7 +84,7 @@ func d3ACalculateQHat(u Digits, v Digits, j int64, n int64) (DoubleDigit, Digit)
 	return qHat, rHat
 }
 
-func d3BTestAndCorrectQHat(b, qHat DoubleDigit, rHat Digit, v Digits, r Digits, j int64, n int64) (DoubleDigit, Digit) {
+func d3BTestAndCorrectQHat(b, qHat digits.DoubleDigit, rHat digits.Digit, v digits.Digits, r digits.Digits, j int64, n int64) (digits.DoubleDigit, digits.Digit) {
 	vLast := v.DigitAt(uint(n - 1))
 
 	vNMinus2 := v.DigitAt(uint(n - 2))
@@ -93,7 +95,7 @@ func d3BTestAndCorrectQHat(b, qHat DoubleDigit, rHat Digit, v Digits, r Digits, 
 
 	// “Now test if qHat ≥ b or qHat * v[n−2] > b*rHat + u[j+n−2]; i”
 	if qHat.IsGreaterThanOrEqual(b) || qHatMulVMinus2.IsGreaterThan(bMulRHatPlusUJPlusNMinus2) {
-		qHat, _ = qHat.Subtract(OneAsDigit().AsDoubleDigit())
+		qHat, _ = qHat.Subtract(digits.OneAsDigit().AsDoubleDigit())
 		rHat, _ = rHat.Add(vLast)
 
 		if rHat.IsLessThan128(b) {
@@ -104,7 +106,7 @@ func d3BTestAndCorrectQHat(b, qHat DoubleDigit, rHat Digit, v Digits, r Digits, 
 			bMulRHatPlusUJPlusNMinus2, _ = bMulRHat.AddDigit(uJPlusNMinus2)
 
 			if qHat.IsGreaterThanOrEqual(b) || qHatMulVMinus2.IsGreaterThan(bMulRHatPlusUJPlusNMinus2) {
-				qHat, _ = qHat.Subtract(OneAsDigit().AsDoubleDigit())
+				qHat, _ = qHat.Subtract(digits.OneAsDigit().AsDoubleDigit())
 				rHat, _ = rHat.Add(vLast)
 			}
 		}
@@ -113,7 +115,7 @@ func d3BTestAndCorrectQHat(b, qHat DoubleDigit, rHat Digit, v Digits, r Digits, 
 	return qHat, rHat
 }
 
-func d4MultiplyAndSubtract(j int64, n int64, u *Digits, v Digits, qHat DoubleDigit) bool {
+func d4MultiplyAndSubtract(j int64, n int64, u *digits.Digits, v digits.Digits, qHat digits.DoubleDigit) bool {
 	ujToJPlusN := u.TakeDigits(uint(j), uint(j)+uint(n))
 	vMulQHat := v.MultiplyDoubleDigit(qHat)
 	ujToJPlusNMinusVMulQHat, borrowed := ujToJPlusN.Trim().SubtractUnderflow(vMulQHat.Trim())
@@ -123,9 +125,9 @@ func d4MultiplyAndSubtract(j int64, n int64, u *Digits, v Digits, qHat DoubleDig
 	return borrowed
 }
 
-func d6AddBack(j int64, qHat DoubleDigit, u *Digits, v Digits) {
-	borrow := Zero()
-	qHat, borrow = qHat.Subtract(OneAsDigit().AsDoubleDigit()) // Decrement q̂
+func d6AddBack(j int64, qHat digits.DoubleDigit, u *digits.Digits, v digits.Digits) {
+	borrow := digits.Zero()                                           // TODO empty
+	qHat, borrow = qHat.Subtract(digits.OneAsDigit().AsDoubleDigit()) // Decrement q̂
 
 	ujToJPlusNMinusVMulQHat := u.TakeDigits(uint(j), uint(j)+uint(v.Length())) // TODO rename
 	ujToJPlusNMinusVMulQHat = ujToJPlusNMinusVMulQHat.Add(v)                   // Add back V to U_j:j+n
@@ -137,7 +139,7 @@ func d6AddBack(j int64, qHat DoubleDigit, u *Digits, v Digits) {
 	}
 }
 
-func divModByDonaldKnuthsTAOCPv2(u Digits, v Digits) (Digits, Digits) {
+func divModByDonaldKnuthsTAOCPv2(u digits.Digits, v digits.Digits) (digits.Digits, digits.Digits) {
 	// Input:
 	// - U = (u_0, u_1, ..., u_{m+n-1})  ⟶ Dividend (m+n digits)
 	// - V = (v_0, v_1, ..., v_{n-1})     ⟶ Divisor (n digits, v_{n-1} ≠ 0)
@@ -150,12 +152,12 @@ func divModByDonaldKnuthsTAOCPv2(u Digits, v Digits) (Digits, Digits) {
 	m := int64(u.Length()) - n
 
 	// Base b (e.g., 10 or 2^64)
-	b := DoubleDigitOf(1, 0)
+	b := digits.DoubleDigitOf(1, 0)
 
 	// D1. [Normalize.]
 	u, v, d := d1Normalize(uint64(n), b, u, v)
 
-	q := Zero().AsDigits()
+	q := digits.Zero().AsDigits()
 
 	// D2. [Initialize j.]
 	for j := m; j >= 0; j-- {
@@ -184,14 +186,14 @@ func divModByDonaldKnuthsTAOCPv2(u Digits, v Digits) (Digits, Digits) {
 	return q.Trim(), remainder.Trim()
 }
 
-func divideByDigit(a Digits, b Digit) (Digits, Digits) {
+func divideByDigit(a digits.Digits, b digits.Digit) (digits.Digits, digits.Digits) {
 	q, r := divideByWord(a, b)
-	return q, MakeDigitsOfDigit(r)
+	return q, digits.MakeDigitsOfDigit(r)
 }
 
-func divideByWord(dividend Digits, divisor Digit) (Digits, Digit) {
-	quotient := Empty()
-	remainder := Zero().AsDoubleDigit()
+func divideByWord(dividend digits.Digits, divisor digits.Digit) (digits.Digits, digits.Digit) {
+	quotient := digits.Empty()
+	remainder := digits.Zero().AsDoubleDigit()
 
 	for i := int64(dividend.Length()) - 1; i >= 0; i-- {
 		di := dividend.DigitAt(uint(i))
@@ -206,21 +208,21 @@ func divideByWord(dividend Digits, divisor Digit) (Digits, Digit) {
 	return quotient.Trim(), remainder.Low()
 }
 
-func divideByDoubleDigit(dividend Digits, divisor DoubleDigit) (Digits, DoubleDigit) {
+func divideByDoubleDigit(dividend digits.Digits, divisor digits.DoubleDigit) (digits.Digits, digits.DoubleDigit) {
 	if divisor.IsZero() {
 		panic("Division by zero")
 	}
 
 	if dividend.IsZero() {
-		return Zero().AsDigits(), Zero().AsDoubleDigit()
+		return digits.Zero().AsDigits(), digits.Zero().AsDoubleDigit()
 	}
 
 	if divisor.IsOne() {
-		return dividend, Zero().AsDoubleDigit()
+		return dividend, digits.Zero().AsDoubleDigit()
 	}
 
 	if dividend.Compare(divisor.AsDigits()) < 0 {
-		return Zero().AsDigits(), dividend.AsDoubleDigit()
+		return digits.Zero().AsDigits(), dividend.AsDoubleDigit()
 	}
 
 	if divisor.High() == 0 {
@@ -233,9 +235,9 @@ func divideByDoubleDigit(dividend Digits, divisor DoubleDigit) (Digits, DoubleDi
 	return quotient.Trim(), remainder
 }
 
-func divideByDoubleWord(dividend Digits, divisor DoubleDigit) (Digits, DoubleDigit) {
-	quotient := Empty()
-	remainder := Zero().AsDigits()
+func divideByDoubleWord(dividend digits.Digits, divisor digits.DoubleDigit) (digits.Digits, digits.DoubleDigit) {
+	quotient := digits.Empty()
+	remainder := digits.Zero().AsDigits()
 
 	for i := int64(dividend.Length()) - 2; i >= 0; i-- {
 		di := dividend.DoubleDigitAt(uint(i))
