+++ divmod/shift.go
@@ -1,7 +1,9 @@
 package divmod
 
+import "github.com/borisskert/go-biginteger/digits"
+
 func ShiftRightUint64Array(a []uint64, n uint64) []uint64 {
-	shifted := Wrap(a).RightShiftBits(n)
+	shifted := digits.Wrap(a).RightShiftBits(n)
 	shifted.NormalizeInPlace()
 
 	return shifted.Trim().Array()
