+++ divmod/subtract.go
@@ -1,7 +1,9 @@
 package divmod
 
+import "github.com/borisskert/go-biginteger/digits"
+
 func Subtract(a, b []uint64) []uint64 {
-	result, _ := Wrap(a).Subtract(Wrap(b))
+	result, _ := digits.Wrap(a).Subtract(digits.Wrap(b))
 	result.NormalizeInPlace()
 
 	return result.Array()
